% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ranges.R
\name{extract_table_ranges}
\alias{extract_table_ranges}
\title{Extract Table Ranges}
\usage{
extract_table_ranges(x)
}
\arguments{
\item{x}{Path to an ODS file containing tables to be extracted.}
}
\value{
A dataframe with three columns: table_name, sheet_name, and cell_range.
        The table_name column contains the names of the tables, sheet_name contains
        the names of the sheets, and cell_range contains the cell ranges of the tables.
}
\description{
Extracts table spans from an ODS file.
}
