% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_tables.R
\name{read_all_tables}
\alias{read_all_tables}
\title{Read all tables from an OpenDocument Spreadsheet (ODS) file, with the option to include or exclude common metadata sheets}
\usage{
read_all_tables(file, exclude_meta = TRUE)
}
\arguments{
\item{file}{Path to the ODS file containing the tables.}

\item{exclude_meta}{Logical indicating whether to exclude metadata tables. Default is TRUE.}
}
\value{
A list of dataframes, each representing a table from the ODS file.
}
\description{
Read all tables from an OpenDocument Spreadsheet (ODS) file, with the option to include or exclude common metadata sheets
}
