% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_table.R
\name{read_ods_table}
\alias{read_ods_table}
\title{Read the data in from a specified table in an OpenDocument Spreadsheet (ODS) file.}
\usage{
read_ods_table(file, tbl_name)
}
\arguments{
\item{file}{Path to the ODS file.}

\item{tbl_name}{Name of the table to be extracted.}
}
\value{
A dataframe of the dats from the specified.
}
\description{
Read the data in from a specified table in an OpenDocument Spreadsheet (ODS) file.
}
